Add list command to lvg and lvol modules
This patch introduces an ability to list the corresponding LVM
entities for lvg and lvol modules which are volume groups and
logical volumes respectively.

New option 'cmd' allow 'list' command to be performed. It returns
a dict where the keys are names of volume groups or logical volumes
depending on which module 'list' command was executed.

This change doesn't modify the behavior of already existing options.

Eg.:
lvol: cmd=list
lvg: cmd=list

In addition to that, it could supply sizes in different units which
is toggled by 'units' option.

Eg.:
lvol: cmd=list units=g

returns the dict of logical volumes where their sizes are in
gigabytes.